+++ apps/xi.front/components/TestComponents/Select/index.tsx
@@ -1,0 +1,18 @@
+import React, { useState } from 'react';
+import { Select } from 'pkg.inputs.select';
+
+export const TestSelect = (staticData: any) => {
+  const [value, setValue] = useState('');
+
+  const changeValue = (newVal: string) => {
+    setValue(newVal);
+  };
+
+  const testData = {
+    ...staticData,
+    value,
+    changeValue,
+  };
+
+  return <Select {...testData} />;
+};
